+++ src/org/cristianlima/controller/MenuDistribuidoresController.java
@@ -18,6 +18,7 @@ import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
 import javafx.fxml.Initializable;
 import javafx.scene.control.Button;
+import javafx.scene.control.ButtonType;
 import javafx.scene.control.TableColumn;
 import javafx.scene.control.TableView;
 import javafx.scene.control.TextField;
@@ -26,6 +27,7 @@ import org.cristianlima.dao.Conexion;
 import org.cristianlima.dto.DistribuidorDTO;
 import org.cristianlima.model.Distribuidor;
 import org.cristianlima.system.Main;
+import org.cristianlima.utils.SuperKinalAlert;
 
 /**
  * FXML Controller class
@@ -72,8 +74,11 @@ public class MenuDistribuidoresController implements Initializable {
             stage.formDistribuidorControllerView(2);
         } else if (event.getSource() == btnEliminar) {
             int disId = ((Distribuidor) tblDistribuidores.getSelectionModel().getSelectedItem()).getDistribuidorId();
+            if(SuperKinalAlert.getInstance().mostrarAlertaConfirmacion(405).get() == ButtonType.OK){
             eliminarDistribuidor(disId);
             cargarLista();
+              
+            }
         } else if (event.getSource() == btnBuscar) {
             tblDistribuidores.getItems().clear();
             if (tfBuscar.getText().isEmpty()) {
@@ -149,6 +154,7 @@ public class MenuDistribuidoresController implements Initializable {
             statement.setInt(1, disId);
             statement.execute();
         } catch (SQLException e) {
+            SuperKinalAlert.getInstance().mostrarAlertaInfo(402);
             System.out.println(e.getMessage());
         } finally {
             try {
